fix creation on nc17
